Removed circular dependencies #2986
