+++ sechub-wrapper-prepare/src/main/java/com/mercedesbenz/sechub/wrapper/prepare/factory/PrepareWrapperPojoFactory.java
@@ -1,23 +1,14 @@
 // SPDX-License-Identifier: MIT
 package com.mercedesbenz.sechub.wrapper.prepare.factory;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
 
 import com.mercedesbenz.sechub.commons.TextFileWriter;
 import com.mercedesbenz.sechub.commons.model.RemoteCredentialContainerFactory;
-import com.mercedesbenz.sechub.commons.pds.PDSUserMessageSupport;
-import com.mercedesbenz.sechub.wrapper.prepare.cli.PrepareWrapperEnvironment;
 
 @Component
 public class PrepareWrapperPojoFactory {
-
-    @Autowired
-    PrepareWrapperEnvironment environment;
-
-    @Autowired
-    TextFileWriter writer;
 
     @Bean
     RemoteCredentialContainerFactory createRemoteCredentialFactory() {
@@ -29,9 +20,4 @@ public class PrepareWrapperPojoFactory {
         return new TextFileWriter();
     }
 
-    @Bean
-    PDSUserMessageSupport createUserMessageSupport() {
-        String userMessageFolder = environment.getPdsUserMessagesFolder();
-        return new PDSUserMessageSupport(userMessageFolder, writer);
-    }
 }
